+++ application.php
@@ -76,6 +76,7 @@ check_session_variables();                              // check session variabl
 
         JOM.new_job.categories.nonce = <?php echo generate_json_javascript_values( '/categories/load', 0, session_id(), $config['SALT'], $config['HASH_ALG'] ); ?>;
         JOM.new_job.issues.nonce     = <?php echo generate_json_javascript_values( '/categories/load', 0, session_id(), $config['SALT'], $config['HASH_ALG'] ); ?>;
+        JOM.new_job.nonce            = <?php echo generate_json_javascript_values( '/job/new',         0, session_id(), $config['SALT'], $config['HASH_ALG'] ); ?>;
         JOM.new_job.get_categories();
 
         jom_init('<?php echo $_SESSION['user']['settings']['i18n']['dateformat']; ?>');