+++ js/categories.js
@@ -1,63 +1,64 @@
 function Categories() {
 
+    var THAT = this;
     /*
        Variable: level
        Needed to select right table, level or sublevel
     */
-    this.level          = undefined;
+    THAT.level          = undefined;
 
     /*
        Variable: parent_id
        Needed if want to get parented subcategories
     */
-    this.parent_id      = undefined;
+    THAT.parent_id      = undefined;
 
     /*
        Variable: categories
        JSON data object
     */
-    this.categories     = undefined;
+    THAT.categories     = undefined;
 
     /*
        Variable: level
        Element container of the categories (starting with SELECT tag
     */
-    this.jq_pointer     = undefined;
+    THAT.jq_pointer     = undefined;
 
     /*
        Variable: nonce
        JSON object that contains nonce and timestamp fields
     */
-    this.nonce          = new Object();
+    THAT.nonce          = new Object();
 
-    this.load = function(level, parent_id) {
+    THAT.load = function(level, parent_id) {
 
         var req = '';
-        if ( this.level     === undefined ) this.level = 1;
+        if ( THAT.level     === undefined ) THAT.level = 1;
 
-        if ( this.parent_id === undefined ) parent_qs = ''
-        else                                parent_qs = '&p=' + this.parent_id;
+        if ( THAT.parent_id === undefined ) parent_qs = ''
+        else                                parent_qs = '&p=' + THAT.parent_id;
 
         $.ajax({
             url:      'ard.php',
-            data:     'd=cat&r=lod&n=' + this.nonce.nonce + '&t=' + this.nonce.timestamp + '&l=' + this.level + parent_qs,
+            data:     'd=cat&r=lod&n=' + THAT.nonce.nonce + '&t=' + THAT.nonce.timestamp + '&l=' + THAT.level + parent_qs,
             type:     'GET',
 			dataType: 'JSON'
         })
         .done(function(data){
-
+            ;   // nothing to do here...
         });
     }
 
-    this.add = function(level, parent_level, subject, description) {
+    THAT.add = function(level, parent_level, subject, description) {
     }
 
-    this.remove = function(level, id) {
+    THAT.remove = function(level, id) {
     }
 
-    this.apply_changes_to_server = function(level, id) {
+    THAT.apply_changes_to_server = function(level, id) {
     }
 
-    this.get_list = function(level) {
+    THAT.get_list = function(level) {
     }
 }