+++ CHANGES.md
@@ -4,9 +4,13 @@ Version 0.3
 TODOS:
 - better implement datepicker (changing date in the input fields does not replect datepicker widget)
 
+
+
+
 11.04.2013
 ----------
 - fixed new-economy message position: now is readable
+- befor sending new job data to server, modal form buttons are disabled, fields disappear, and a "saving" message appears
 
 10.04.2013
 ----------